backend/requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
pillow==8.4.0
psycopg2==2.9.1
